requirements.txt
altgraph==0.17.4
asttokens==2.4.1
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
ImageHash==4.3.1
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
MouseInfo==0.1.3
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
parso==0.8.3
pefile==2023.2.7
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.17.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
pytweening==1.0.7
PyWavelets==1.5.0
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.1
scipy==1.11.4
screeninfo==0.8.1
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
wcwidth==0.2.11
